Allow setting a sub-field of a vcell if BME is on
Add a new bit-manipulation crate feature that turns on the new set_field
method.  Generate a BME address with the BFI opcode.  Limited to setting
a number of bits X, 1 <= X <= 16 -- that is, up to half of the word at a
time.

This does *no* safety checking on the start-bit or bit-count fields; if
they are too big their values simply get truncated.