Fix bugs: check SetField range, set BB addr right
Panic if SetField parameters are out of range, since the compiler can
prove that this will never happen at runtime (since the addresses are
always hardcoded from svd2rust's generated code), so we don't pay the
cost.  And this is required for safety.

Also shift the bit-index field in the bitband address; add a big comment
explaining why it's safe and necessary, but basically the bottom two
bits always have to be zero, because a bitband address needs to be an
aligned u32 pointer.