Add tests for bitbanding and BME
Make sure that all valid addresses for bitbanding generate aligned u32
pointers.  Make sure that the addresses on the edge of the invalid range
all panic.

Make sure that all valid SetField combinations for BME generate the same
bits in the resulting address in places that don't include the first-bit
or bit-count values.  Make sure that invalid SetField combinations
panic.